Adds field for optional related links, Fixes #3

+++ src/FileUtils.js
@@ -120,7 +120,8 @@ const addNote = (uri, filedata) => {
             } else {
                 dbtitle.ensureIndex({ fieldName: 'title', unique: true });
                 let imageData = filedata.imageData;
-                let imageFileName = uuid();
+                let imageFileName = '';
+                if (imageData) imageFileName = uuid();
                 filedata.imageData = imageFileName;
                 dbtitle.update(
                     { title: filedata.title },
@@ -155,6 +156,9 @@ const addNote = (uri, filedata) => {
 };
 const deleteImage = (uri, imageFileName) => {
     return new Promise((resolve, reject) => {
+        if (imageFileName === '') {
+            return resolve();
+        }
         fs.unlink(path.join(uri, imageFileName))
             .then(() => {
                 resolve();
@@ -212,6 +216,9 @@ const deleteNote = (uri, title) => {
 };
 const saveImage = (uri, imageData, imageFileName) => {
     return new Promise((resolve, reject) => {
+        if (imageFileName === '') {
+            return resolve();
+        }
         fs.writeFile(path.join(uri, imageFileName), imageData)
             .then(() => resolve())
             .catch(() => reject());
